HypixelBazaarViewerDISCORD/requirements.txt
discord.py
requests
natsort
